requirements.txt
ccxt>=2.7.5
pandas>=1.5.0
numpy>=1.22.0
matplotlib>=3.5.0
ta>=0.10.0
python-dotenv>=0.19.0
plotly>=5.8.0
scikit-learn>=1.0.0
tabulate>=0.8.0
rich>=12.0.0
pytz>=2022.1
python-telegram-bot>=13.0
requests>=2.28.0
bs4>=0.0.1
joblib>=1.1.0
fastapi>=0.95.0  # Для будущего веб-интерфейса
uvicorn>=0.22.0  # Для запуска FastAPI приложения